feat: streamlit menubar
#19
