+++ tools/.streamlit/config.toml
@@ -3,3 +3,6 @@ headless = true
 port = 8501
 enableCORS = false
 address = "0.0.0.0"
+
+[theme]
+base = "light"
