+++ tools/views/about_us.py
@@ -1,0 +1,6 @@
+
+import streamlit as st
+
+def app():
+    st.title('About us')
+    st.write('Welcome to the About us!')
