+++ tools/views/dashboard.py
@@ -1,0 +1,6 @@
+# pages/home.py
+import streamlit as st
+
+def app():
+    st.title('Dashboard')
+    st.write('Welcome to the Dashboard!')
