+++ tools/views/home.py
@@ -1,0 +1,6 @@
+# pages/home.py
+import streamlit as st
+
+def app():
+    st.title('Home Page')
+    st.write('Welcome to the Home Page!')
